requirements.txt
appdirs==1.4.4
Babel==2.14.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
frozendict==2.4.0
html5lib==1.1
idna==3.6
kiwisolver==1.4.5
lxml==5.0.0
matplotlib==3.8.2
multitasking==0.0.11
numpy==1.26.3
packaging==23.2
pandas==2.1.4
peewee==3.17.0
pillow==10.2.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
tkcalendar==1.6.1
tzdata==2023.4
urllib3==2.1.0
webencodings==0.5.1
wheel==0.41.2
yfinance==0.2.33
